Refactoring and layout edit screen.
