+++ Joey/Resources/layout/EditGroupedTimeEntryItem.axml
@@ -3,24 +3,24 @@
     android:orientation="horizontal"
     android:layout_width="fill_parent"
     android:layout_height="70dp">
-    <View 
+    <View
         android:id="@+id/GroupedEditTimeEntryItemTimeColorView"
-    	android:layout_width="5dp"
-    	android:layout_height="fill_parent"
-    	android:background="#ffffff" />
+        android:layout_width="5dp"
+        android:layout_height="fill_parent"
+        android:background="#ffffff" />
     <ImageView
         android:layout_width="wrap_content"
         android:layout_height="fill_parent"
         android:src="@drawable/icClock"
-        android:layout_marginLeft="26dp"
+        android:layout_marginLeft="16dp"
         android:layout_marginRight="5dp" />
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
         android:textColor="#313131"
         android:textStyle="normal"
-		android:fontFamily="sans-serif-light" 
-		android:textSize="16sp"
+        android:fontFamily="sans-serif-light"
+        android:textSize="16sp"
         android:id="@+id/GroupedEditTimeEntryItemTimePeriodTextView"
         android:layout_gravity="center_horizontal"
         android:gravity="center" />
@@ -29,7 +29,7 @@
         android:layout_height="match_parent"
         android:textColor="#313131"
         android:textSize="16sp"
-        android:layout_marginRight="26dp"
+        android:layout_marginRight="16dp"
         android:id="@+id/GroupedEditTimeEntryItemDurationTextView"
         android:gravity="right|center_vertical" />
 </LinearLayout>
