+++ Joey/Resources/layout/EditTimeEntryFragment.axml
@@ -3,6 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
     <LinearLayout
+    	android:background="@drawable/BgArrowsFull"
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
