--- Joey/Resources/layout/GroupedEditTimeEntryFragment.axml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:background="#ffffff">
-    <LinearLayout
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:background="@drawable/BgArrowsFull">
-        <android.support.v7.widget.Toolbar
-            android:id="@+id/GroupedEditTimeEntryFragmentToolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:minHeight="?attr/actionBarSize"
-            android:background="@null" />
-        <LinearLayout
-            android:orientation="vertical"
-            android:minWidth="25px"
-            android:minHeight="25px"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingRight="32dp"
-            android:paddingLeft="32dp"
-            android:layout_marginTop="15dp">
-            <Toggl.Joey.UI.Views.TogglField
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:id="@+id/GroupedEditTimeEntryFragmentProject" />
-            <Toggl.Joey.UI.Views.TogglField
-                android:layout_width="match_parent"
-				android:layout_height="match_parent"
-                android:id="@+id/GroupedEditTimeEntryFragmentDescription"
-                android:layout_marginTop="20dp" />
-            <Toggl.Joey.UI.Views.EditTimeEntryTagsBit
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:id="@+id/GroupedEditTimeEntryFragmentTags"
-                android:layout_marginTop="20dp" />
-            <CheckBox
-                android:id="@+id/GroupedEditTimeEntryFragmentBillableCheckBox"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:text="@string/CurrentTimeEntryEditBillableUnchecked"
-                android:textSize="16sp"
-                android:textColor="@color/BillableTextColor"
-                android:gravity="center_vertical"
-                android:button="@null"
-                android:layout_marginTop="24dp"
-                android:drawableLeft="@drawable/IcBillableGreen"
-                android:drawablePadding="5dp" />
-        </LinearLayout>
-    </LinearLayout>
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_weight="2">
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/recyclerView"
-            android:layout_weight="2"
-            android:background="#f8f8f8"
-            android:layout_width="match_parent"
-            android:layout_height="fill_parent" />
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="8dp"
-            android:background="@drawable/DropShadow" />
-    </FrameLayout>
-</LinearLayout>
